# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlackDiamondClub Gambling Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  recoveryQuestion: string;
  recoveryAnswer: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface RecoverInput {
  username: string;
  recoveryAnswer: string;
  /** @minLength 6 */
  newPassword: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Wallet {
  userId: number;
  depositBalance: number;
  withdrawalBalance: number;
  totalBalance?: number;
  currency: string;
}

export interface DepositAddress {
  address: string;
  network: string;
  currency: string;
  /** Base64 encoded QR code image */
  qrCode: string;
  minDeposit: number;
}

export interface DepositInput {
  amount: number;
  /** @nullable */
  txHash?: string | null;
  /**
     * Base64 encoded screenshot image of the payment
     * @nullable
     */
  screenshotBase64?: string | null;
}

export interface WithdrawalInput {
  amount: number;
  walletAddress: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  game_win: 'game_win',
  game_loss: 'game_loss',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Transaction {
  id: number;
  userId: number;
  type: TransactionType;
  amount: number;
  status: TransactionStatus;
  /** @nullable */
  txHash?: string | null;
  /** @nullable */
  screenshotUrl?: string | null;
  createdAt: string;
  /** @nullable */
  gameType?: string | null;
}

export interface TransactionList {
  transactions: Transaction[];
  total: number;
}

export type BetInputGameType = typeof BetInputGameType[keyof typeof BetInputGameType];


export const BetInputGameType = {
  aviator: 'aviator',
  slots777: 'slots777',
  rocketrush: 'rocketrush',
  coinflip: 'coinflip',
  spinsprint: 'spinsprint',
  dicedash: 'dicedash',
  towerclimb: 'towerclimb',
  luckywheel: 'luckywheel',
  numberblast: 'numberblast',
  gemdrop: 'gemdrop',
} as const;

export interface BetInput {
  gameType: BetInputGameType;
  amount: number;
  /**
     * Optional prediction for games that need it (e.g. heads/tails for coinflip)
     * @nullable
     */
  prediction?: string | null;
}

export interface CashoutInput {
  sessionId: number;
  currentMultiplier: number;
}

export interface BetResult {
  sessionId: number;
  won: boolean;
  multiplier: number;
  betAmount: number;
  winAmount: number;
  newDepositBalance: number;
  newWithdrawalBalance: number;
  /** @nullable */
  crashPoint?: number | null;
  message: string;
}

export type FakePlayerAction = typeof FakePlayerAction[keyof typeof FakePlayerAction];


export const FakePlayerAction = {
  entered: 'entered',
  won: 'won',
  lost: 'lost',
} as const;

export interface FakePlayer {
  username: string;
  amount: number;
  /** @nullable */
  multiplier?: number | null;
  action: FakePlayerAction;
  time: string;
}

export interface GameStats {
  gameType: string;
  activePlayers: number;
  recentActivity: FakePlayer[];
  bigWins: FakePlayer[];
}

export interface LeaderboardEntry {
  rank: number;
  username: string;
  totalWinnings: number;
  gamesPlayed: number;
  winRate: number;
  isOnline: boolean;
}

export interface LeaderboardList {
  leaders: LeaderboardEntry[];
  updatedAt: string;
}

export interface UserProfile {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
  recoveryQuestion: string;
}

export interface ProfileUpdate {
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  currentPassword?: string | null;
  /** @nullable */
  newPassword?: string | null;
}

export interface ProfileStats {
  totalBets: number;
  totalWins: number;
  totalLosses: number;
  winRate: number;
  biggestWin: number;
  totalDeposited: number;
  totalWithdrawn: number;
  /** @nullable */
  favoriteGame?: string | null;
}

export type ListTransactionsParams = {
type?: ListTransactionsType;
limit?: number;
offset?: number;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  deposit: 'deposit',
  withdrawal: 'withdrawal',
  game_win: 'game_win',
  game_loss: 'game_loss',
} as const;

export type GetGameStatsParams = {
gameType?: string;
};

export type GetLeaderboardParams = {
limit?: number;
};