# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BlackDiamondClub Gambling Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current logged-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/recover:
    post:
      operationId: recoverAccount
      tags: [auth]
      summary: Verify recovery question and reset password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoverInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Wrong answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /games/bet:
    post:
      operationId: placeBet
      tags: [games]
      summary: Place a bet on a game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BetInput"
      responses:
        "200":
          description: Bet result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResult"
        "400":
          description: Insufficient funds or invalid bet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /games/cashout:
    post:
      operationId: cashout
      tags: [games]
      summary: Cash out during an active crash game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CashoutInput"
      responses:
        "200":
          description: Cashout result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResult"
        "400":
          description: No active bet or game already crashed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /games/stats:
    get:
      operationId: getGameStats
      tags: [games]
      summary: Get recent game stats and fake live activity
      parameters:
        - name: gameType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Game stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameStats"
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - username
        - createdAt
    RegisterInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        recoveryQuestion:
          type: string
        recoveryAnswer:
          type: string
      required:
        - username
        - password
        - recoveryQuestion
        - recoveryAnswer
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    RecoverInput:
      type: object
      properties:
        username:
          type: string
        recoveryAnswer:
          type: string
        newPassword:
          type: string
          minLength: 6
      required:
        - username
        - recoveryAnswer
        - newPassword
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
...
    BetInput:
      type: object
      properties:
        gameType:
          type: string
          enum: [aviator, slots777, rocketrush, coinflip, spinsprint, dicedash, towerclimb, luckywheel, numberblast, gemdrop]
        amount:
          type: number
        prediction:
          type: ["string", "null"]
          description: Optional prediction for games that need it (e.g. heads/tails for coinflip)
      required:
        - gameType
        - amount
    CashoutInput:
      type: object
      properties:
        sessionId:
          type: integer
        currentMultiplier:
          type: number
      required:
        - sessionId
        - currentMultiplier
    BetResult:
      type: object
      properties:
        sessionId:
          type: integer
        won:
          type: boolean
        multiplier:
          type: number
        betAmount:
          type: number
        winAmount:
          type: number
        newDepositBalance:
          type: number
        newWithdrawalBalance:
          type: number
        crashPoint:
          type: ["number", "null"]
        message:
          type: string
      required:
        - sessionId
        - won
        - multiplier
        - betAmount
        - winAmount
        - newDepositBalance
        - newWithdrawalBalance
        - message
    FakePlayer:
      type: object
      properties:
        username:
          type: string
        amount:
          type: number
        multiplier:
          type: ["number", "null"]
        action:
          type: string
          enum: [entered, won, lost]
        time:
          type: string
      required:
        - username
        - amount
        - action
        - time
    GameStats:
      type: object
      properties:
        gameType:
          type: string
        activePlayers:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/FakePlayer"
        bigWins:
          type: array
          items:
            $ref: "#/components/schemas/FakePlayer"
      required:
        - gameType
        - activePlayers
        - recentActivity
        - bigWins
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
        recoveryQuestion:
          type: string
      required:
        - id
        - username
        - createdAt
        - recoveryQuestion
    ProfileUpdate:
      type: object
      properties:
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        currentPassword:
          type: ["string", "null"]
        newPassword:
          type: ["string", "null"]
...