# lib/api-zod/src/generated/types/cashoutInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BlackDiamondClub Gambling Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CashoutInput {
  sessionId: number;
  currentMultiplier: number;
}